refactor(music): move player components to ui/components/musicplayer
