+++ app/src/main/java/io/github/aedev/flow/ui/FlowApp.kt
@@ -44,13 +44,13 @@ import io.github.aedev.flow.player.GlobalPlayerState
 import io.github.aedev.flow.player.SleepTimerManager
 import io.github.aedev.flow.ui.components.DonationPromptHost
 import io.github.aedev.flow.ui.components.FloatingBottomNavBar
-import io.github.aedev.flow.ui.components.MusicPlayerBottomSheet
-import io.github.aedev.flow.ui.components.MusicPlayerSheetState
-import io.github.aedev.flow.ui.components.PersistentMiniMusicPlayer
+import io.github.aedev.flow.ui.components.musicplayer.MusicPlayerBottomSheet
+import io.github.aedev.flow.ui.components.musicplayer.MusicPlayerSheetState
+import io.github.aedev.flow.ui.components.musicplayer.PersistentMiniMusicPlayer
 import io.github.aedev.flow.ui.components.PlayerSheetValue
 import io.github.aedev.flow.ui.components.SleepTimerSheet
 import io.github.aedev.flow.ui.components.layout.topbar.ProvideFlowGlobalActions
-import io.github.aedev.flow.ui.components.rememberMusicPlayerSheetState
+import io.github.aedev.flow.ui.components.musicplayer.rememberMusicPlayerSheetState
 import io.github.aedev.flow.ui.components.rememberPlayerDraggableState
 import io.github.aedev.flow.ui.screens.home.HomeViewModel
 import io.github.aedev.flow.ui.screens.music.EnhancedMusicPlayerScreen
