+++ app/src/main/java/io/github/aedev/flow/ui/components/musicplayer/MusicPaletteColors.kt
@@ -1,4 +1,4 @@
-package io.github.aedev.flow.ui.screens.music.player
+package io.github.aedev.flow.ui.components.musicplayer
 
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.tween
