+++ app/src/main/java/io/github/aedev/flow/ui/components/musicplayer/PlayerTopBar.kt
@@ -1,4 +1,4 @@
-package io.github.aedev.flow.ui.screens.music.player
+package io.github.aedev.flow.ui.components.musicplayer
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
