+++ app/src/main/java/io/github/aedev/flow/ui/screens/settings/PlayerAppearanceScreen.kt
@@ -49,8 +49,8 @@ import io.github.aedev.flow.data.local.SliderStyle
 import io.github.aedev.flow.data.local.resolveSeekbarHorizontalPaddingDp
 import io.github.aedev.flow.ui.components.layout.topbar.FlowTopBar
 import io.github.aedev.flow.ui.components.rememberFlowSheetState
-import io.github.aedev.flow.ui.screens.music.player.components.PlayerSliderTrack
-import io.github.aedev.flow.ui.screens.music.player.components.SquigglySlider
+import io.github.aedev.flow.ui.components.musicplayer.PlayerSliderTrack
+import io.github.aedev.flow.ui.components.musicplayer.SquigglySlider
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
