+++ app/src/main/java/io/github/aedev/flow/ui/tv/music/TvLyricsPanel.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.res.stringResource
 import io.github.aedev.flow.R
 import io.github.aedev.flow.ui.screens.music.MusicPlayerUiState
 import io.github.aedev.flow.ui.screens.music.MusicTrack
-import io.github.aedev.flow.ui.screens.music.player.InlineLyricsPanel
+import io.github.aedev.flow.ui.components.musicplayer.InlineLyricsPanel
 import io.github.aedev.flow.ui.tv.components.TvSidePanel
 import io.github.aedev.flow.ui.tv.focus.tvInitialFocus
