+++ app/src/main/java/io/github/aedev/flow/ui/tv/music/TvMusicNowPlayingScreen.kt
@@ -60,9 +60,9 @@ import io.github.aedev.flow.data.local.PlayerPreferences
 import io.github.aedev.flow.player.EnhancedMusicPlayerManager
 import io.github.aedev.flow.player.RepeatMode
 import io.github.aedev.flow.ui.screens.music.MusicPlayerViewModel
-import io.github.aedev.flow.ui.screens.music.player.PlayerBackground
-import io.github.aedev.flow.ui.screens.music.player.PlayerProgressSlider
-import io.github.aedev.flow.ui.screens.music.player.rememberMusicPalette
+import io.github.aedev.flow.ui.components.musicplayer.PlayerBackground
+import io.github.aedev.flow.ui.components.musicplayer.PlayerProgressSlider
+import io.github.aedev.flow.ui.components.musicplayer.rememberMusicPalette
 import io.github.aedev.flow.ui.tv.components.TvIconButton
 import io.github.aedev.flow.ui.tv.components.TvIconButtonColors
 import io.github.aedev.flow.ui.tv.input.TvPlayerAction
